Remove @shopify/polaris, @shopify/polaris-icons and react-dom from cli-kit
Rewrite the GraphiQL page templates as plain HTML strings instead of
rendering Polaris React components with renderToStaticMarkup at request
time. The generated HTML is byte-identical to the previous output.
