+++ packages/cli-kit/package.json
@@ -107,8 +107,6 @@
     "@graphql-typed-document-node/core": "3.2.0",
     "@iarna/toml": "2.2.5",
     "@oclif/core": "4.8.3",
-    "@shopify/polaris": "12.27.0",
-    "@shopify/polaris-icons": "8.11.1",
     "@shopify/toml-patch": "0.3.0",
     "@opentelemetry/api": "1.9.1",
     "@opentelemetry/core": "2.8.0",
@@ -155,7 +153,6 @@
     "pathe": "1.1.2",
     "react": "19.2.4",
     "semver": "7.8.4",
-    "react-dom": "19.2.4",
     "stacktracey": "2.2.0",
     "strip-ansi": "7.2.0",
     "supports-hyperlinks": "3.2.0",
@@ -169,7 +166,6 @@
     "@types/gradient-string": "^1.1.2",
     "@types/lodash": "4.17.24",
     "@types/react": "^19.0.0",
-    "@types/react-dom": "^19.0.0",
     "@types/semver": "^7.5.2",
     "@types/which": "3.0.4",
     "@vitest/coverage-istanbul": "^3.2.7",
